+++ src/helpers/index.js
@@ -45,4 +45,4 @@ export {
 export AlignBlock from '@plone/volto/helpers/AlignBlock/AlignBlock';
 
 export Helmet from './Helmet/Helmet';
-export { difference } from '@plone/volto/helpers/Utils/Utils';
+export * from '@plone/volto/helpers/Utils/Utils';